rag_module/pyproject.toml
[tool.poetry]
name = "rag-module"
version = "0.1.0"
description = ""
authors = ["Aiswarya-s2413 <aiswaryaakku9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.118.0"
uvicorn = "^0.37.0"
python-multipart = "^0.0.20"
python-dotenv = "^1.1.1"
pydantic = "^2.11.10"
pypdf2 = "^3.0.1"
requests = "^2.32.5"
sentence-transformers = "^5.1.1"
pinecone-client = "^5.0.0"
langchain = "^0.3.27"
langchain-text-splitters = "^0.3.11"
langchain-core = "^0.3.28"
langchain-community = "^0.3.25"
openai = "^1.54.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
boto3 = "^1.35.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
